requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.21
aiosqlite>=0.19.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
loguru>=0.7.0
email-validator>=2.0.0
psutil>=5.9.0
# 新增依赖
opencv-python>=4.5.5
numpy>=1.22.0
websocket-client>=1.4.0
websockets>=11.0.0
pyyaml>=6.0 